EMBED-903 Demonstrate context data usage
A configuration view has been added that allows a user to show/hide other views and to override the keys for the embed views which were hard coded. A new home page has been added that shows some blurb about what the Kitchen Sink does. Neither the configuration view or the home view can be hidden. The example is contrived as normally the configuration would be restricted to some kind of administrator.

The App component has been split into two components, App and KitchenSink. This has been done as the KitchenSink needs access to context data. App is now a very simple component that wraps the KitchenSink component in the ExtensionProvider. This allows the KitchenSink to access the context without the ugliness (at least in my opinion) that would need to happen if the code remained in the App component.