webpack 5 stricter about accessing environment variables from process.env.
With this change, the webpack EnvironmentPlugin is now used to access those environment variables that are actually required rather than all of the environment variables defined in .env. The added advantage is that the EnvironmentPlugin gives an explicit error message about missing variables when the development server is started.

In addition, as the README did not define the POSTS_SERVER_URL environment variable (which is required), a default value is provided.